// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const RoleName = {
  CASHIER: 'CASHIER',
  INVENTORY: 'INVENTORY'
} as const

export type RoleName = (typeof RoleName)[keyof typeof RoleName]


export const GlobalRole = {
  SUPER_ADMIN: 'SUPER_ADMIN',
  ADMIN: 'ADMIN',
  USER: 'USER'
} as const

export type GlobalRole = (typeof GlobalRole)[keyof typeof GlobalRole]


export const StockMovementType = {
  STOCK_IN: 'STOCK_IN',
  STOCK_OUT: 'STOCK_OUT',
  ADJUSTMENT: 'ADJUSTMENT',
  SALE: 'SALE',
  RETURN: 'RETURN'
} as const

export type StockMovementType = (typeof StockMovementType)[keyof typeof StockMovementType]


export const PaymentMethod = {
  CASH: 'CASH',
  CREDIT: 'CREDIT',
  MIXED: 'MIXED'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const SaleStatus = {
  DRAFT: 'DRAFT',
  COMPLETED: 'COMPLETED',
  VOIDED: 'VOIDED'
} as const

export type SaleStatus = (typeof SaleStatus)[keyof typeof SaleStatus]


export const NotificationType = {
  LOW_STOCK: 'LOW_STOCK',
  OVERDUE_CREDIT: 'OVERDUE_CREDIT'
} as const

export type NotificationType = (typeof NotificationType)[keyof typeof NotificationType]